move from url to strings, add a Slugify description:
