--- url.go
@@ -1,15 +1,0 @@
-package swiss
-
-import (
-	"regexp"
-	"strings"
-)
-
-func Slugify(s string) string {
-	s = strings.ToLower(s)
-	reg := regexp.MustCompile("[^a-z0-9]+")
-	s = reg.ReplaceAllString(s, "-")
-	s = strings.Trim(s, "-")
-
-	return s
-}
